+++ db/migrate/20140409174157_add_email_sent_to_dataset_user.rb
@@ -1,0 +1,5 @@
+class AddEmailSentToDatasetUser < ActiveRecord::Migration
+  def change
+    add_column :dataset_users, :email_sent, :boolean, null: false, default: false
+  end
+end