+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Minor wording changes on About page and Home page
 - Documentation pages for tools and datasets now include the filename as part of the title
 - Variable popups are now hidden when the `Esc` key is pressed
+- Users are notified by email when a dataset editor approves their file access request
 - **Gem Changes**
   - Updated to rails 4.1.0
   - Updated to contour 2.5.0.beta1